Rewriting Chapter 2 Finger Print cache, use slab allocator to manage memory, use linked list | binary search tree | skip list as different search algorithm.

+++ c_cpp/0002/allocator.h
@@ -92,7 +92,7 @@ extern void align_to_pow2(uint64_t *size, uint64_t pow2);
 extern void* allocator_alloc(allocator_t *allocator);
 extern void allocator_free(allocator_t *allocator, void *block);
 
-extern size_t allocator_slab_reclaim(allocator_t *allocator, 
+extern int32_t allocator_slab_reclaim(allocator_t *allocator, 
         void (*relocate) (const void *source, void * dst, size_t block_size, void *user_data),
         void *user_data);
